Replace VisualizationThemeSwitch
