+++ context/app/static/js/components/Detail/VisualizationThemeSwitch/index.js
@@ -1,0 +1,3 @@
+import VisualizationThemeSwitch from './VisualizationThemeSwitch';
+
+export default VisualizationThemeSwitch;
