+++ context/app/static/js/components/Detail/VisualizationThemeSwitch/style.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+import ToggleButton from '@material-ui/lab/ToggleButton';
+
+const StyledToggleButton = styled(ToggleButton)`
+  display: inline-block;
+  background-color: #ffffff;
+  border: 0px;
+  svg {
+    vertical-align: middle;
+  }
+`;
+
+export { StyledToggleButton };
